+++ src/path.mli
@@ -111,3 +111,5 @@ val rm_rf : t -> unit
 
 (** Changes the extension of the filename (or adds an extension if there was none) *)
 val change_extension : ext:string -> t -> t
+
+val pp : t Fmt.t